# routes/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "This is a sample YoYo Gifts Server.  You can find out more about     Yoyo Gifts at [https://yoyo-gift-qa.web.app](https://yoyo-gift-qa.web.app) or on [irc.freenode.net, #swagger](https://yoyo-gift-qa.web.app/irc/).      For this sample, you can use the api key `special-key` to test the authorization     filters.",
    "version": "1.0.0",
    "title": "YO YO Gifts",
    "termsOfService": "https://yoyo-gift-qa.web.app/terms/",
    "contact": {
      "email": "apiteam@yoyo-gift.com"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
  "host": "yoyo-gift-qa.web.app",
...
  "schemes": "https",
  "paths": {
...
    "/profileDetails/getById/{_id}": {
      "get": {
        "tags": ["user"],
        "summary": "Find user by Id",
        "description": "find details of an user by user id which is coming through params",
        "operationId": "createUsersWithListInput",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "_id",
            "in": "path",
            "description": "ID of user that needs to be fetched",
            "required": true,
            "schema": {
              "$ref": "#/definitions/FindUserById"
            }
          }
        ],
        "responses": {
          "404": {
            "description": " 'userDetail not found with id"
          },
          "200": {
            "description": " 'Fetching UserDetail Successfully",
            "userDetail": "userDetail"
          }
        }
      }
    },
...
    "/profileDetails/updateyoyobalance/{_id}": {
      "put": {
        "tags": ["user"],
        "summary": "yoyo balance by Id",
        "description": "find details of an user by user id which is coming through params",
        "operationId": "createUsersWithListInput",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "description": "Authorization token required",
            "required": true,
            "type": "string"
          },
          {
            "name": "isAdmin",
            "in": "header",
            "description": "A flag to check type of user",
            "default": false,
            "required": true,
            "type": "boolean"
          },
          {
            "name": "orderdetails",
            "in": "body",
            "description": "ID of user that needs to be fetched",
            "required": true,
            "schema": {
              "$ref": "#/definitions/yoyobalance"
            }
          },
          {
            "name": "_id",
            "in": "path",
            "description": "ID of user that needs to be fetched",
            "required": true,
            "schema": {
              "$ref": "#/definitions/yoyobalance"
            }
          }
        ],
        "responses": {
          "404": {
            "description": " 'userDetail not found with id"
          },
          "200": {
            "description": " 'Fetching UserDetail Successfully",
            "userDetail": "userDetail"
          }
        }
      }
    },
    "/profileDetails/updateName/{_id}": {
      "put": {
        "tags": ["user"],
        "summary": "yoyo balance by Id",
        "description": "find details of an user by user id which is coming through params",
        "operationId": "createUsersWithListInput",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "description": "Authorization token required",
            "required": true,
            "type": "string"
          },
          {
            "name": "isAdmin",
            "in": "header",
            "description": "A flag to check type of user",
            "default": false,
            "required": true,
            "type": "boolean"
          },
          {
            "name": "Name",
            "in": "body",
            "description": "ID of user that needs to be fetched",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Name"
            }
          },
          {
            "name": "_id",
            "in": "path",
            "description": "ID of user that needs to be fetched",
            "required": true,
            "schema": {
              "$ref": "#/definitions/yoyobalance"
            }
          }
        ],
        "responses": {
          "404": {
            "description": " 'userDetail not found with id"
          },
          "200": {
            "description": " 'Fetching UserDetail Successfully",
            "userDetail": "userDetail"
          }
        }
      }
    },
...
    "/profileDetails/updateById/{_id}": {
      "put": {
        "tags": ["user"],
        "summary": "Find by Id and update the details of user",
        "description": "Find by Id and update the details of user",
        "operationId": "updateById",
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "description": "Authorization token required",
            "required": true,
            "type": "string"
          },
          {
            "name": "isAdmin",
            "in": "header",
            "description": "A flag to check type of user",
            "default": false,
            "required": true,
            "type": "boolean"
          },
          {
            "name": "_id",
            "in": "path",
            "description": "ID of user that needs to be fetched",
            "required": true,
            "schema": {
              "$ref": "#/definitions/FindUserById"
            }
          },
          {
            "in": "body",
            "name": "body",
            "description": "Details of new user who is trying to signup.",
            "required": true,
            "schema": {
              "$ref": "#/definitions/UserUpdate"
            }
          }
        ],
        "responses": {
          "400": {
            "description": "UserDetail not found with id"
          },
          "200": {
            "description": "Fetching UserDetail Successfully",
            "userDetail": "userDetail"
          }
        }
      }
    }
  },
  "securityDefinitions": {
    "petstore_auth": {
      "type": "oauth2",
      "authorizationUrl": "http://petstore.swagger.io/oauth/dialog",
      "flow": "implicit",
      "scopes": {
        "write:pets": "modify pets in your account",
        "read:pets": "read your pets"
      }
    },
    "api_key": {
      "type": "apiKey",
      "name": "api_key",
      "in": "header"
    }
  },
  "definitions": {
...
    "Name": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string",
          "example": "Test"
        }
      }
    },
...
    "yoyobalance": {
      "type": "object",
      "properties": {
        "yoyobalance": {
          "type": "string",
          "example": "323"
        }
      }
    },
...
    "UserUpdate": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string"
        },
        "email": {
          "type": "string"
        }
      }
    },
...
  },
...
}